--- libs/router.js
@@ -1,15 +1,0 @@
-
-
-var express = require('express');
-var router = express.Router();
-
-router.use(function timeLog(req, res, next) {
-  console.log('Time: ', Date.now());
-  next();
-});
-
-router.get('/', function(req, res) {
-  res.send('test');
-});
-
-module.exports = router;
