+++ package.json
@@ -4,7 +4,7 @@
   "description": "software for call centers",
   "main": "app.js",
   "scripts": {
-    "test": "node app.js"
+    "test": "node ./bin/www"
   },
   "author": "Unixson",
   "license": "MIT",
@@ -12,8 +12,11 @@
     "body-parser": "^1.18.2",
     "config.json": "0.0.4",
     "connect": "^3.6.5",
+    "cookie-parser": "^1.4.3",
     "express": "^4.16.2",
     "mongodb": "^3.0.2",
+    "pug": "^2.0.0-rc.4",
+    "serve-favicon": "~2.4.5",
     "simcom": "^0.3.0",
     "socket.io": "^2.0.4",
     "winston": "^2.4.0"
